Normalize PubChem compound references
PubChem is an ambiguous prefix that could refer to one of the several semantic spaces. This PR updates them to use a more standardized prefix based on the Bioregistry's standard. While PRs are normally made to the -edit file, it looks like this one is out of sync, so I made the changes in all of the artifacts. Would be worth looking into properly reconciling all of the artifacts in this repo as well.